Partial fix for bug #691172: The tiff resolution tags are floating point values and need to be passed as such to TIFFSetField.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@10925 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/base/gdevtifs.c
@@ -245,8 +245,8 @@ int tiff_set_fields_for_printer(gx_device_printer *pdev, TIFF *tif)
     TIFFSetField(tif, TIFFTAG_PLANARCONFIG, PLANARCONFIG_CONTIG);
 
     TIFFSetField(tif, TIFFTAG_RESOLUTIONUNIT, RESUNIT_INCH);
-    TIFFSetField(tif, TIFFTAG_XRESOLUTION, pdev->x_pixels_per_inch);
-    TIFFSetField(tif, TIFFTAG_YRESOLUTION, pdev->y_pixels_per_inch);
+    TIFFSetField(tif, TIFFTAG_XRESOLUTION, (float)pdev->x_pixels_per_inch);
+    TIFFSetField(tif, TIFFTAG_YRESOLUTION, (float)pdev->y_pixels_per_inch);
 
     {
 	char revs[10];
